% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_issues.R
\name{get_repos_from_open_issues}
\alias{get_repos_from_open_issues}
\title{derive target repos from all open issues for a contributions-like repository}
\usage{
get_repos_from_open_issues(owner, repo)
}
\arguments{
\item{owner}{character(1) github.com repo owner}

\item{repo}{character(1) repo name}
}
\value{
A vector of URLs supplied by contributors, identifying the URL of their contribution's repo.
}
\description{
derive target repos from all open issues for a contributions-like repository
}
\note{
The python code was written by instructing perplexity.ai on Nov 18 2025.
}
\examples{
rs = get_repos_from_open_issues("bioconductor", "contributions")
head(rs)
}
