% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymods4issues.R
\name{pymods4issues}
\alias{pymods4issues}
\title{obtain modules for interrogating and summarizing a github issues stream}
\usage{
pymods4issues()
}
\description{
obtain modules for interrogating and summarizing a github issues stream
}
\examples{
pp = pymods4issues()  # wants GITHUB_TOKEN and OPENAI_API_KEY
names(pp)
}
