% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_submission.R
\name{summarize_submission}
\alias{summarize_submission}
\title{summarize the repo underlying a submission issue}
\usage{
summarize_submission(url)
}
\arguments{
\item{url}{character(1) with correct https URL for contribution repo}
}
\value{
A list with components `summary` (text string with summary, and `url`
(the repo URL string.  See inst/python/summ1.py for hardcoded openai parameters
}
\description{
summarize the repo underlying a submission issue
}
\note{
At present we have hardcoded the LLM to use as gpt-4. At most 2000 words from DESCRIPTION
and at most 4000 words from README are provided for a brief prompt asking for technical summary.

The python code was written by instructing perplexity.ai on Nov 18 2025.
}
\examples{
t1 = summarize_submission("https://github.com/bioconductor/AnVILVRS")
cat(strwrap(t1, 75), sep="\n")
}
